rename to finetune_rag
